backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
PyMySQL==1.1.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
Pillow==10.2.0
